order_service/order/serializers.py
# order_service/order/serializers.py
from rest_framework import serializers
from .models import Order

class OrderSerializer(serializers.ModelSerializer):
    class Meta:
        model = Order
        fields = ('id', 'user_email', 'product_name', 'quantity','price')
