order_service/requirements.txt
Django
djangorestframework
django-cors-headers
celery
pika
psycopg2-binary